update profile page

+++ app/Http/Controllers/PackageController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use App\Models\Package;
+use App\Models\Project;
 use Illuminate\Support\Str;
 
 
@@ -66,8 +67,8 @@ class PackageController extends Controller
         $package->status_upload = $this->readStatus($package->status_upload);
         $package->status_running = $this->readStatus($package->status_running);
 
-        if(!$package->staging == 'yes') {
-            $package->stagin = 'Ada 0000_staging';
+        if($package->staging == 'yes') {
+            $package->staging = 'Ada 0000_staging';
         } else if($package->staging == 'no') {
             $package->staging = 'Tidak ada 0000_staging';
         } else {
@@ -78,9 +79,32 @@ class PackageController extends Controller
             $package->keterangan  = '-';
         }
 
+        $projects = Project::where('package_id', $package->id)
+                        ->with(['jobs.job_steps'])
+                        ->orderBy('created_at', 'asc')
+                        ->paginate(10);
+
+        // dd($projects);
+
+        foreach ($projects as $project) {
+            if($project->staging == 'yes') {
+                $project->staging = 'Ada';
+            } else if($project->staging == 'no') {
+                $project->staging = 'Tidak Ada';
+            } else {
+                $project->staging = 'Tidak ada';
+            }
+
+            $project->status_migrasi = $this->readStatus($project->status_migrasi);
+            // $project->staging = $this->readStatus($project->staging);
+            $project->status_upload = $this->readStatus($project->status_upload);
+            $project->status_running = $this->readStatus($project->status_running);
+            $project->status_import = $this->readStatus($project->status_import);
+        }
+
         return view('pages.package.detail.index', [
             'package' => $package,
-            'projects' => []
+            'projects' => $projects
         ]);
     }
 
@@ -163,10 +187,17 @@ class PackageController extends Controller
     }
 
     public function delete_package($id) {
-        $package = Package::find($id);
+        $package = Package::with(['projects'])->find($id);
 
         if(!$package) {
             return abort(404);
+        }
+
+
+        if(count($package->projects) > 0) {
+            return back()->withErrors([
+                'message' => "Package $package->name tidak bisa dihapus karena memiliki project, untuk menghapus item package hapus terlebih dahulu project-project yang terkait dengan package ini"
+            ]);
         }
         
         $package->delete();
